Add test coverage for changes and fix existing unit test
